Fix #17 - Prevent long beeps
It turns out some printers have issues generating a beep in excess of 4 seconds (Prusa has had an open defect for a year and a half) which can lead to printer resets. While it would be nice if those were fixed, for the time being, let's just set a cap on how long the beep can last.

In addition, I tweaked the help string for the Trigger Beep button now that #24 has gone in